+++ openai-java-core/src/test/kotlin/com/openai/services/async/ImageServiceAsyncTest.kt
@@ -26,7 +26,7 @@ internal class ImageServiceAsyncTest {
         val imagesResponseFuture =
             imageServiceAsync.createVariation(
                 ImageCreateVariationParams.builder()
-                    .image("some content".toByteArray())
+                    .image("some content".byteInputStream())
                     .model(ImageModel.DALL_E_2)
                     .n(1L)
                     .responseFormat(ImageCreateVariationParams.ResponseFormat.URL)
@@ -51,9 +51,9 @@ internal class ImageServiceAsyncTest {
         val imagesResponseFuture =
             imageServiceAsync.edit(
                 ImageEditParams.builder()
-                    .image("some content".toByteArray())
+                    .image("some content".byteInputStream())
                     .prompt("A cute baby sea otter wearing a beret")
-                    .mask("some content".toByteArray())
+                    .mask("some content".byteInputStream())
                     .model(ImageModel.DALL_E_2)
                     .n(1L)
                     .quality(ImageEditParams.Quality.HIGH)