+++ openai-java-core/src/test/kotlin/com/openai/services/blocking/uploads/PartServiceTest.kt
@@ -24,7 +24,7 @@ internal class PartServiceTest {
             partService.create(
                 PartCreateParams.builder()
                     .uploadId("upload_abc123")
-                    .data("some content".toByteArray())
+                    .data("some content".byteInputStream())
                     .build()
             )